Introduce separate module for mini-backend.
