+++ libs/wire-subsystems/src/Wire/MiniBackend.hs
@@ -1,0 +1,313 @@
+module Wire.MiniBackend where
+
+import Data.Default (Default (def))
+import Data.Domain
+import Data.Id
+import Data.LanguageCodes (ISO639_1 (EN))
+import Data.LegalHold (UserLegalHoldStatus (UserLegalHoldDisabled))
+import Data.Map.Lazy qualified as LM
+import Data.Map.Strict qualified as M
+import Data.Proxy
+import Data.Qualified
+import Data.Set qualified as S
+import Data.Time
+import Data.Type.Equality
+import Imports
+import Polysemy
+import Polysemy.Error
+import Polysemy.Input
+import Polysemy.State
+import Servant.Client.Core
+import Test.QuickCheck
+import Type.Reflection
+import Wire.API.Federation.API
+import Wire.API.Federation.Component
+import Wire.API.Federation.Error
+import Wire.API.Team.Member
+import Wire.API.User hiding (DeleteUser)
+import Wire.DeleteQueue
+import Wire.FederationAPIAccess
+import Wire.FederationAPIAccess.Interpreter as FI
+import Wire.GalleyAPIAccess
+import Wire.InternalEvent
+import Wire.Sem.Concurrency
+import Wire.Sem.Concurrency.Sequential
+import Wire.Sem.Now
+import Wire.StoredUser
+import Wire.UserStore
+import Wire.UserSubsystem
+import Wire.UserSubsystem.Interpreter
+
+newtype PendingStoredUser = PendingStoredUser StoredUser
+  deriving (Show, Eq)
+
+instance Arbitrary PendingStoredUser where
+  arbitrary = do
+    user <- arbitrary
+    pure $ PendingStoredUser (user {status = Just PendingInvitation})
+
+newtype NotPendingStoredUser = NotPendingStoredUser StoredUser
+  deriving (Show, Eq)
+
+instance Arbitrary NotPendingStoredUser where
+  arbitrary = do
+    user <- arbitrary
+    notPendingStatus <- elements (Nothing : map Just [Active, Suspended, Deleted, Ephemeral])
+    pure $ NotPendingStoredUser (user {status = notPendingStatus})
+
+type GetUserProfileEffects =
+  [ UserSubsystem,
+    GalleyAPIAccess,
+    UserStore,
+    DeleteQueue,
+    State [InternalNotification],
+    Now,
+    Input UserSubsystemConfig,
+    FederationAPIAccess MiniFederationMonad,
+    Concurrency 'Unsafe,
+    Error FederationError
+  ]
+
+-- | a type representing the state of a single backend
+data MiniBackend = MkMiniBackend
+  { -- | this is morally the same as the users stored in the actual backend
+    --   invariant: for each key, the user.id and the key are the same
+    users :: Set StoredUser
+  }
+
+instance Default MiniBackend where
+  def = MkMiniBackend {users = mempty}
+
+-- | represents an entire federated, stateful world of backends
+newtype MiniFederation = MkMiniFederation
+  { -- | represents the state of the backends, mapped from their domains
+    backends :: Map Domain MiniBackend
+  }
+
+data MiniContext = MkMiniContext
+  { -- | the domain that is receiving the request
+    ownDomain :: Domain
+  }
+
+newtype MiniFederationMonad comp a = MkMiniFederationMonad
+  {unMiniFederation :: Sem [Input MiniContext, State MiniFederation] a}
+  deriving newtype (Functor, Applicative, Monad)
+
+instance RunClient (MiniFederationMonad comp) where
+  runRequestAcceptStatus _acceptableStatuses _req = error "MiniFederation does not support servant client"
+  throwClientError _err = error "MiniFederation does not support servant client"
+
+data SubsystemOperationList where
+  TNil :: SubsystemOperationList
+  (:::) :: Typeable a => (Component, Text, a) -> SubsystemOperationList -> SubsystemOperationList
+
+infixr 5 :::
+
+lookupSubsystemOperation ::
+  Typeable a =>
+  -- | The type to compare to
+  (Component, Text, Proxy a) ->
+  -- | what to return when none of the types match
+  a ->
+  -- | the types to try
+  SubsystemOperationList ->
+  a
+lookupSubsystemOperation goal@(goalComp, goalRoute, Proxy @goalType) a = \case
+  TNil -> a
+  (comp, route, client) ::: xs -> case eqTypeRep (typeRep @goalType) (typeOf client) of
+    Just HRefl | comp == goalComp && route == goalRoute -> client
+    _ -> lookupSubsystemOperation goal a xs
+
+instance FederationMonad MiniFederationMonad where
+  fedClientWithProxy (Proxy @name) (Proxy @api) (_ :: Proxy (MiniFederationMonad comp)) =
+    -- TODO(mangoiv): checking the type of the Client alone is not enough; we also want to check that
+    -- the route is correct; this has yet to be implemented by storing an existential of the route
+    -- or, more easily a `Text` that represents it
+    lookupSubsystemOperation
+      (componentVal @comp, nameVal @name, Proxy @(Client (MiniFederationMonad comp) api))
+      do
+        error
+          "The testsuite has evaluated a tuple of component, route and client that is\
+          \ not covered by the MiniFederation implementation of FederationMonad"
+      do
+        (Brig, "get-users-by-ids", miniGetUsersByIds)
+          ::: (Brig, "get-user-by-id", miniGetUserById)
+          ::: TNil
+
+miniLocale :: Locale
+miniLocale =
+  Locale
+    { lLanguage = Language EN,
+      lCountry = Nothing
+    }
+
+-- | runs a stateful backend, returns the state and puts it back into the
+--   federated global state
+runOnOwnBackend :: Sem '[Input Domain, State MiniBackend] a -> MiniFederationMonad comp a
+runOnOwnBackend = MkMiniFederationMonad . runOnOwnBackend' . subsume_
+
+-- | Runs action in the context of a single backend, without access to others.
+runOnOwnBackend' ::
+  (Member (Input MiniContext) r, Member (State MiniFederation) r) =>
+  Sem (Input Domain ': State MiniBackend ': r) a ->
+  Sem r a
+runOnOwnBackend' act = do
+  ownDomain <- inputs (.ownDomain)
+  ownBackend <-
+    fromMaybe (error "tried to lookup domain that is not part of the backends' state")
+      <$> gets (M.lookup ownDomain . backends)
+  (newBackend, res) <- runState ownBackend $ runInputConst ownDomain act
+  modify (\minifed -> minifed {backends = M.insert ownDomain newBackend (minifed.backends)})
+  pure res
+
+miniGetAllProfiles ::
+  (Member (Input Domain) r, Member (State MiniBackend) r) =>
+  Sem r [UserProfile]
+miniGetAllProfiles = do
+  users <- gets (.users)
+  dom <- input
+  pure $
+    map
+      (\u -> mkUserProfileWithEmail Nothing (mkUserFromStored dom miniLocale u) UserLegalHoldDisabled)
+      (S.toList users)
+
+miniGetUsersByIds :: [UserId] -> MiniFederationMonad 'Brig [UserProfile]
+miniGetUsersByIds userIds = runOnOwnBackend do
+  usersById :: LM.Map UserId UserProfile <-
+    M.fromList . map (\user -> (user.profileQualifiedId.qUnqualified, user)) <$> miniGetAllProfiles
+  pure $ mapMaybe (flip M.lookup usersById) userIds
+
+miniGetUserById :: UserId -> MiniFederationMonad 'Brig (Maybe UserProfile)
+miniGetUserById uid =
+  runOnOwnBackend $
+    find (\u -> u.profileQualifiedId.qUnqualified == uid) <$> miniGetAllProfiles
+
+runMiniFederation :: Domain -> Map Domain MiniBackend -> MiniFederationMonad c a -> a
+runMiniFederation ownDomain backends =
+  run
+    . evalState MkMiniFederation {backends = backends}
+    . runInputConst MkMiniContext {ownDomain = ownDomain}
+    . unMiniFederation
+
+interpretNowConst ::
+  UTCTime ->
+  Sem (Now : r) a ->
+  Sem r a
+interpretNowConst time = interpret \case
+  Wire.Sem.Now.Get -> pure time
+
+runFederationStack ::
+  [StoredUser] ->
+  Map Domain MiniBackend ->
+  Maybe TeamMember ->
+  UserSubsystemConfig ->
+  Sem GetUserProfileEffects a ->
+  a
+runFederationStack allLocalUsers fedBackends =
+  runFederationStackWithUnavailableBackends allLocalUsers fedBackends mempty
+
+runFederationStackWithUnavailableBackends ::
+  [StoredUser] ->
+  -- | the available backend
+  Map Domain MiniBackend ->
+  -- | the down backend
+  Map Domain MiniBackend ->
+  Maybe TeamMember ->
+  UserSubsystemConfig ->
+  Sem GetUserProfileEffects a ->
+  a
+runFederationStackWithUnavailableBackends allLocalUsers availableBackends unavailableBackends teamMember cfg =
+  let unsafeError e = error $ "Unexpected error: " <> displayException e
+   in either unsafeError Imports.id
+        . runFederationStackWithUnavailableBackendsEither
+          allLocalUsers
+          availableBackends
+          unavailableBackends
+          teamMember
+          cfg
+
+runFederationStackWithUnavailableBackendsEither ::
+  [StoredUser] ->
+  -- | the available backend
+  Map Domain MiniBackend ->
+  -- | the down backend
+  Map Domain MiniBackend ->
+  Maybe TeamMember ->
+  UserSubsystemConfig ->
+  Sem GetUserProfileEffects a ->
+  Either FederationError a
+runFederationStackWithUnavailableBackendsEither allLocalUsers availableBackends unavailableBackends teamMember cfg =
+  run
+    . runError
+    . sequentiallyPerformConcurrency
+    . miniFederationAPIAccess availableBackends unavailableBackends
+    . runInputConst cfg
+    . interpretNowConst (UTCTime (ModifiedJulianDay 0) 0)
+    . evalState []
+    . inMemoryDeleteQueueInterpreter
+    . staticUserStoreInterpreter allLocalUsers
+    . miniGalleyAPIAccess teamMember
+    . runUserSubsystem cfg
+
+runNoFederationStack ::
+  [StoredUser] ->
+  Maybe TeamMember ->
+  UserSubsystemConfig ->
+  Sem GetUserProfileEffects a ->
+  a
+runNoFederationStack allUsers teamMember cfg =
+  run
+    . runErrorUnsafe
+    . sequentiallyPerformConcurrency
+    . emptyFederationAPIAcesss
+    . runInputConst cfg
+    . interpretNowConst (UTCTime (ModifiedJulianDay 0) 0)
+    . evalState []
+    . inMemoryDeleteQueueInterpreter
+    . staticUserStoreInterpreter allUsers
+    . miniGalleyAPIAccess teamMember
+    . runUserSubsystem cfg
+
+inMemoryDeleteQueueInterpreter :: Member (State [InternalNotification]) r => InterpreterFor DeleteQueue r
+inMemoryDeleteQueueInterpreter = interpret $ \case
+  EnqueueUserDeletion uid -> modify (\l -> DeleteUser uid : l)
+  EnqueueClientDeletion cid uid mConnId -> modify (\l -> DeleteClient cid uid mConnId : l)
+  EnqueueServiceDeletion pid sid -> modify (\l -> DeleteService pid sid : l)
+
+runErrorUnsafe :: Exception e => InterpreterFor (Error e) r
+runErrorUnsafe action = do
+  res <- runError action
+  case res of
+    Left e -> error $ "Unexpected error: " <> displayException e
+    Right x -> pure x
+
+emptyFederationAPIAcesss :: InterpreterFor (FederationAPIAccess MiniFederationMonad) r
+emptyFederationAPIAcesss = interpret $ \case
+  _ -> error "uninterpreted effect: FederationAPIAccess"
+
+miniFederationAPIAccess ::
+  forall a r.
+  Map Domain MiniBackend ->
+  Map Domain MiniBackend ->
+  Sem (FederationAPIAccess MiniFederationMonad : r) a ->
+  Sem r a
+miniFederationAPIAccess online _offline = do
+  let runner :: FederatedActionRunner MiniFederationMonad r
+      runner domain rpc = pure . Right $ runMiniFederation domain online rpc
+  interpret \case
+    RunFederatedEither remote rpc ->
+      if isJust (M.lookup (qDomain $ tUntagged remote) online)
+        then FI.runFederatedEither runner remote rpc
+        else pure $ Left do FederationUnexpectedError "RunFederatedEither"
+    RunFederatedConcurrently _remotes _rpc -> error "unimplemented: RunFederatedConcurrently"
+    RunFederatedBucketed _domain _rpc -> error "unimplemented: RunFederatedBucketed"
+    IsFederationConfigured -> pure True
+
+staticUserStoreInterpreter :: [StoredUser] -> InterpreterFor UserStore r
+staticUserStoreInterpreter allUsers = interpret $ \case
+  GetUser uid -> pure $ find (\user -> user.id == uid) allUsers
+
+miniGalleyAPIAccess :: Maybe TeamMember -> InterpreterFor GalleyAPIAccess r
+miniGalleyAPIAccess member = interpret $ \case
+  GetTeamMember _ _ -> pure member
+  _ -> error "uninterpreted effect: GalleyAPIAccess"
